+++ CameraControl.Devices/Others/YiCamera.cs
@@ -69,7 +69,7 @@ namespace CameraControl.Devices.Others
             {
                 IsoNumber = new PropertyValue<long> {Available = false};
                 FNumber = new PropertyValue<long> {Available = false};
-                ExposureCompensation = new PropertyValue<int> {Available = false};
+                ExposureCompensation = new PropertyValue<long> {Available = false};
                 FocusMode = new PropertyValue<long> {Available = false};
                 ShutterSpeed = new PropertyValue<long> {Available = false};
                 WhiteBalance = new PropertyValue<long> {Available = false};
